Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-control-expected.txt


FAIL pointer updates (control test) assert_equals: error events on c expected 1 but got 0
